amendmant made at L29 to let the computer know it has to do floating point division

+++ csharp-prep/Prep4/Program.cs
@@ -26,7 +26,7 @@ class Program
         }
 
         int terms = numbers.Count;
-        float average = sum / terms;
+        float average = ((float) sum) / terms;
 
         int max = 0;
         foreach (int number in numbers) {
